requirements.txt
django==5.0.6
djangorestframework==3.15.1
psycopg2-binary==2.9.9
ipython==8.24.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
python-dotenv==1.0.1